requirements.txt
gunicorn
python-decouple
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==44.0.3
Django==5.2
django-allauth==65.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
Markdown==3.8
mysqlclient==2.2.7
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
